fixed a problem related to largefile support
under certain circumstances tombupnp was compiled with disabled largefile 
support, while mediatomb was compiled with largefile support

as the result the off_t sizes in tombupnp and mediatomb differed.

it turned out that we used mysql_config --cflags which somehow
declared largefile defines on some systems.

since we set those only for the mediatomb and not for the tombupnp part, 
the off_t sizes differed.

adapted configure.ac and removed workarounds from upnp.h/upnpapi.c


git-svn-id: https://svn.mediatomb.cc/svnroot/mediatomb/trunk/mediatomb@460 f010a08f-7000-0410-b1fc-cb24403df210