completely reworked the custom http headers patch. previously we were going through the linked list each time a HTTP message was sent out -> not cool if you think of performance
so now there is still the global linked list (for ease of use when removing 
and adding headers) and then there is a concatenated copy of the list which
is used in HTTP responses. of course we are wasting a little memory here,
but I don't think it's critical in that case.

one big todo: the SDK virtual directory functions as well as my newly added
custom header functions are not threadsafe. they should only be called once
after the initialization.

currently this is not a problem because we do not change anyhting while the
server is running, but we have to be aware of that.

also moved all custom user headers to the last position in the HTTP message.



git-svn-id: https://svn.mediatomb.cc/svnroot/mediatomb/trunk/mediatomb@498 f010a08f-7000-0410-b1fc-cb24403df210