we are now handling the rpl_realloc case
AC_FUNC_MALLOC and AC_FUNC_REALLOC will redefine malloc and realloc if
they think that we are not building for the GNU C library. This can be
a problem for cross compile environments, so our configure script now has
an option to disable this feature.

However, we also added appropriate rpl_ functions to handle the cases
where malloc is indeed not GNU C compatible.



git-svn-id: https://svn.mediatomb.cc/svnroot/mediatomb/trunk/mediatomb@693 f010a08f-7000-0410-b1fc-cb24403df210