fixed a problem related to taglib/libexif
configure uses $LIBS when it does AC_CHECK_LIB tests

unfortunately some libs that are not required for a particular
checks are added. this can cause problems - when checking for 
-lexif, -ltag was also added - and this required -lz on the NSLU2,
however -lz was not part of the $LIBS variable so the exif check
failed because of libraries that had nothing to do with it!

now saving $LIBS and unsetting the variable during exif tests



git-svn-id: https://svn.mediatomb.cc/svnroot/mediatomb/trunk/mediatomb@727 f010a08f-7000-0410-b1fc-cb24403df210